backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
sqlalchemy>=2.0.36
pydantic>=2.10.0
pydantic-settings>=2.6.0
pydantic[email]>=2.10.0
PyJWT>=2.9.0
passlib>=1.7.4
bcrypt>=4.0.1
python-multipart>=0.0.20
slowapi>=0.1.9
email-validator>=2.0.0
